requirements.txt
amqp==5.3.1
asgiref==3.8.1
asttokens==3.0.0
attrs==25.1.0
backcall==0.2.0
beautifulsoup4==4.13.3
billiard==4.2.1
bleach==6.2.0
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap4==2024.10
cron-descriptor==1.4.5
decorator==5.2.1
defusedxml==0.7.1
Django==5.1.6
django-allauth==65.4.1
django-appconf==1.0.6
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-crispy-bootstrap==0.1.1.1
django-crispy-forms==2.3
django-haystack==3.3.0
django-humanize==0.1.2
django-imagekit==5.0.0
django-redis==5.4.0
django-simpleui==2025.1.13
django-timezone-field==7.1
docker==7.1.0
docopt==0.6.2
executing==2.2.0
fastjsonschema==2.21.1
humanize==4.11.0
idna==3.10
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kombu==5.4.2
Markdown==3.7
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.2
mysqlclient==2.0.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pilkit==3.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pymdown-extensions==10.14.3
pypinyin==0.53.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
Whoosh==2.7.4
yarg==0.1.9
